// front/src/app/models/ucenici.ts
export class ucenici{
    username: string = "";
    password: string = "";
    odgovorNaPitanje: string = "";
    ime: string = "";
    prezime: string = "";
    pol: string = "";
    adresa: string = "";
    telefon: string = "";
    email: string = "";
    profilePictureUrl: string = "";
    tipSkole: string = "";
    razred: number = -1;
    brojOcena: number = 0;
    ukupnaOcena: number = 0;
}